+++ test/unit/rendering.js
@@ -220,4 +220,64 @@
     start();
   });
 
+  asyncTest( 'When no `cursor` option is specified, default cursor is used with color based on OS (black for Mac, white for Windows and Linux)', function(){
+    expect(1);
+    var
+      mac= (/macintosh/i).test(navigator.userAgent),
+      $reel= $('#image').reel({
+      })
+
+    $reel.bind('loaded.test', function(){
+      equal( $(this).css('cursor').replace(' ', ''), 'url(http://code.vostrel.cz/jquery.reel-'+(mac? 'black':'white')+'.cur),move');
+      start();
+    });
+  });
+
+  asyncTest( 'When `cursor: "hand"` is given the legacy grasping hand cursor will be used', function(){
+    expect(2);
+    var
+      $reel= $('#image').reel({
+        cursor: 'hand'
+      })
+
+    $reel.bind('loaded.test', function(){
+      equal( $(this).css('cursor').replace(' ', ''), 'url(http://code.vostrel.cz/jquery.reel-drag.cur),move');
+
+      // Simulate dragging/panning
+      $('html').addClass('reel-panning');
+      equal( $(this).css('cursor').replace(' ', ''), 'url(http://code.vostrel.cz/jquery.reel-drag-down.cur),move');
+      $('html').removeClass('reel-panning');
+
+      start();
+    });
+  });
+
+  asyncTest( 'Otherwise `cursor` option accepts any valid CSS cursor declaration', function(){
+    expect(1);
+    var
+      $reel= $('#image').reel({
+        cursor: 'pointer'
+      })
+
+    $reel.bind('loaded.test', function(){
+      equal( $(this).css('cursor'), 'pointer');
+      start();
+    });
+  });
+
+  asyncTest( 'Instance being preloaded has a "hourglass" cursor to indicate the pending loading', function(){
+    expect(4);
+    var
+      $reel= $('#image').reel({
+      })
+
+    ok( $reel.parent().is('.reel-loading'), 'Is `reel-loading` at the very start');
+    equal ($reel.css('cursor'), 'wait', 'Has the "wait" cursor');
+    $reel.bind('loaded.test', function(){
+      ok( !$reel.parent().is('.reel-loading'), 'No longer has the `reel-loading` class when loaded');
+      ok ($reel.css('cursor') != 'wait', 'Does not have the "wait" cursor after the load');
+      start();
+    });
+  });
+
 })(jQuery);